Разбиение хранилища на два.
